refactor: 컴포넌트 폴더 구조 정리 및 미사용 파일 제거(#184)
- Logo 컴포넌트를 commons로 이동
- ProductCard를 commons/product로 이동
- useIntersectionObserver를 components/hook으로 이동
- DetailFilterButton을 filter 폴더로 이동
- ProductsSection을 product-section 폴더로 이동
- 미사용 파일 제거: SearchIcon, SellerOtherProductList
