+++ src/components/commons/product/ProductCard.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom'
-import type { Product } from '../../types'
-import { ProductThumbnail } from './ProductThumbnail'
-import { ProductInfo } from './ProductInfo'
+import type { Product } from '../../../types'
+import { ProductThumbnail } from '../../product/ProductThumbnail'
+import { ProductInfo } from '../../product/ProductInfo'
 import { ROUTES } from '@src/constants/routes'
 import { getTradeStatus } from '@src/utils/getTradeStatus'
 import { getPetTypeName } from '@src/utils/getPetTypeName'
