+++ src/pages/home/components/filter/DetailFilter.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react'
-import { DetailFilterButton } from '../DetailFilterButton'
+import { DetailFilterButton } from './DetailFilterButton'
 import { ProductStateFilter } from '../../../../components/commons/filters/ProductStateFilter'
 import { PriceFilter } from '../../../../components/commons/filters/PriceFilter'
 import { LocationFilter } from '../../../../components/commons/filters/LocationFilter'
