--- src/pages/product-detail/components/SellerOtherProductList.tsx
@@ -1,11 +1,0 @@
-export default function SellerOtherProductList() {
-  return (
-    <ul className="tablet:grid-cols-3 desktop:grid-cols-4 gap-lg grid grid-cols-2">
-      {/* {sellerOtherProducts?.map((sellerProducts, index) => (
-            <li key={sellerProducts.id}>
-              <ProductCard data-index={index} data={sellerProducts} />
-            </li>
-          ))} */}
-    </ul>
-  )
-}
